(BA-1077) baseapp modularization
