+++ apps/docs/pages/_app.tsx
@@ -1,4 +1,4 @@
-import { QueryClient, QueryClientProvider } from 'react-query'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
 export default function MyApp({ Component, pageProps }: { Component: any; pageProps: object }) {
   const client = new QueryClient()
