--- apps/storybook-app/components/Checkbox/index.tsx
@@ -1,3 +1,0 @@
-import { CheckboxField, ICheckboxFieldProps } from "@baseapp-frontend/design-system-mui";
-
-export const Checkbox = ({...props}: ICheckboxFieldProps) => <CheckboxField {...props} />
