+++ apps/storybook/.storybook/main.ts
@@ -1,9 +1,7 @@
 import type { StorybookConfig } from '@storybook/nextjs'
 
 const config: StorybookConfig = {
-  stories: [
-    '../components/**/stories.@(js|jsx|ts|tsx)',
-  ],
+  stories: ['../components/**/stories.@(js|jsx|ts|tsx)'],
   addons: [
     '@storybook/addon-links',
     '@storybook/addon-essentials',
