+++ apps/storybook/components/Checkbox/index.tsx
@@ -1,0 +1,3 @@
+import { CheckboxField, ICheckboxFieldProps } from '@baseapp-frontend/design-system-mui'
+
+export const Checkbox = ({ ...props }: ICheckboxFieldProps) => <CheckboxField {...props} />
