+++ apps/storybook/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "storybook-app",
+  "name": "storybook",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
